+++ worker/src/RTC/RTCP/FeedbackRtpTransport.cpp
@@ -271,7 +271,7 @@ namespace RTC
 			}
 
 			// 32 bits padding.
-			const size_t padding = (-offset) & 3;
+			const size_t padding = Utils::Byte::PadTo4Bytes(offset) - offset;
 
 			for (size_t i{ 0u }; i < padding; ++i)
 			{
@@ -380,7 +380,7 @@ namespace RTC
 				size += 2u;
 
 				// 32 bits padding.
-				size += (-size) & 3;
+				size = Utils::Byte::PadTo4Bytes(size);
 
 				if (size > maxRtcpPacketLen)
 				{