+++ worker/src/RTC/RTCP/Sdes.cpp
@@ -226,7 +226,7 @@ namespace RTC
 			++offset;
 
 			// 32 bits padding.
-			const size_t padding = (-offset) & 3;
+			const size_t padding = Utils::Byte::PadTo4Bytes(offset) - offset;
 
 			for (size_t i{ 0u }; i < padding; ++i)
 			{