+++ worker/src/handles/BackoffTimerHandle.cpp
@@ -128,7 +128,7 @@ int64_t BackoffTimerHandle::ComputeNextTimeoutMs() const
 			return std::min<int64_t>(timeoutMs, BackoffTimerHandleInterface::MaxTimeoutMs);
 		}
 
-			NO_DEFAULT_GCC();
+			NO_DEFAULT();
 	}
 }